# Fetch messages a bit faster