--- a/src/protocol.rs
+++ b/src/protocol.rs
@@ -1379,3 +1379,68 @@ impl FromByte for Message {
         )
     }
 }
+
+#[cfg(test)]
+mod tests {
+
+    use std::io::Cursor;
+
+    use super::FetchResponse;
+    use codecs::FromByte;
+
+    #[test]
+    fn decode_new_fetch_response() {
+
+        // - one topic
+        // - 2 x message of 10 bytes (0..10)
+        // - 2 x message of 5 bytes (0..5)
+        // - 3 x message of 10 bytes (0..10)
+        // - 1 x message of 5 bytes (0..5) static
+        static FETCH_RESPONSE_RAW_DATA: &'static [u8] = &[
+            0, 0, 0, 3, 0, 0, 0, 1, 0, 13, 116, 101, 115, 116, 95,
+            116, 111, 112, 105, 99, 95, 49, 112, 0, 0, 0, 1, 0, 0, 0,
+            0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 8, 0, 0, 1, 17, 0, 0, 0, 0,
+            0, 0, 0, 0, 0, 0, 0, 24, 211, 120, 76, 139, 0, 0, 255,
+            255, 255, 255, 0, 0, 0, 10, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9,
+            0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 24, 211, 120, 76, 139, 0,
+            0, 255, 255, 255, 255, 0, 0, 0, 10, 0, 1, 2, 3, 4, 5, 6,
+            7, 8, 9, 0, 0, 0, 0, 0, 0, 0, 2, 0, 0, 0, 19, 224, 237,
+            15, 248, 0, 0, 255, 255, 255, 255, 0, 0, 0, 5, 0, 1, 2, 3,
+            4, 0, 0, 0, 0, 0, 0, 0, 3, 0, 0, 0, 19, 224, 237, 15, 248,
+            0, 0, 255, 255, 255, 255, 0, 0, 0, 5, 0, 1, 2, 3, 4, 0, 0,
+            0, 0, 0, 0, 0, 4, 0, 0, 0, 24, 211, 120, 76, 139, 0, 0,
+            255, 255, 255, 255, 0, 0, 0, 10, 0, 1, 2, 3, 4, 5, 6, 7,
+            8, 9, 0, 0, 0, 0, 0, 0, 0, 5, 0, 0, 0, 24, 211, 120, 76,
+            139, 0, 0, 255, 255, 255, 255, 0, 0, 0, 10, 0, 1, 2, 3, 4,
+            5, 6, 7, 8, 9, 0, 0, 0, 0, 0, 0, 0, 6, 0, 0, 0, 24, 211,
+            120, 76, 139, 0, 0, 255, 255, 255, 255, 0, 0, 0, 10, 0, 1,
+            2, 3, 4, 5, 6, 7, 8, 9, 0, 0, 0, 0, 0, 0, 0, 7, 0, 0, 0,
+            19, 224, 237, 15, 248, 0, 0, 255, 255, 255, 255, 0, 0, 0,
+            5, 0, 1, 2, 3, 4];
+
+        let r = FetchResponse::decode_new(&mut Cursor::new(FETCH_RESPONSE_RAW_DATA));
+        let msgs = r.unwrap().into_messages();
+
+        macro_rules! assert_msg {
+            ($msg:expr, $topic:expr, $partition:expr, $msgdata:expr) => {
+                assert_eq!($topic, &$msg.topic[..]);
+                assert_eq!($partition, $msg.partition);
+                assert_eq!($msgdata, &$msg.message[..]);
+            }
+        }
+
+        assert_eq!(8, msgs.len());
+        let zero_to_ten: Vec<u8> = (0..10).collect();
+        assert_msg!(msgs[0], "test_topic_1p", 0, &zero_to_ten[..]);
+        assert_msg!(msgs[1], "test_topic_1p", 0, &zero_to_ten[..]);
+
+        assert_msg!(msgs[2], "test_topic_1p", 0, &zero_to_ten[0..5]);
+        assert_msg!(msgs[3], "test_topic_1p", 0, &zero_to_ten[0..5]);
+
+        assert_msg!(msgs[4], "test_topic_1p", 0, &zero_to_ten[..]);
+        assert_msg!(msgs[5], "test_topic_1p", 0, &zero_to_ten[..]);
+        assert_msg!(msgs[6], "test_topic_1p", 0, &zero_to_ten[..]);
+
+        assert_msg!(msgs[7], "test_topic_1p", 0, &zero_to_ten[0..5]);
+    }
+}